Fix inconsistent formatting issue in convert function when converting from money to varchar (#3337) (#3360)
Description
Currently when converting from money to varchar using convert function with different style values, the result is inconsistent with the formatting. This PR will address this issue by updating the logic to handle these edge cases. The issues that will get addressed in this PR are as follows

Extra space is getting added in the beginning of the result when style parameter is 0 or 2.
When input is 0, then 0 is missing before decimal in the result for style 0 and 2. Also for style 1 the result has $ before decimal instead of 0. For example, when converting value 0 with money datatype to varchar datatype
For style 0, result is  .00
For style 1, result is $.00
For style 2, result is  .0000
When style is 2, and the input has more than 2 digits after decimal, then the result is getting rounded upto 2 digits after decimal, which ideally should be rounded upto 4 digits.

Issues Resolved
BABEL-5462

Authored-by: Rohit Bhagat rohitbgt@amazon.com
Signed-off-by: Rohit Bhagat rohitbgt@amazon.com