Cross-db insert on table with identity throws permission error on inserting in identity column (#3436)
Previously, INSERTs on identity columns would throw permission denied on sequence error if the relation name is referenced as a three part object name and the database name in the qualified object name is not the current database name. To fix this, when we are about to proceed with the identity insert, we will first identify it is a cross database scenario. If it is, we will change the current role as the session user.

Task: BABEL-5585
Signed-off-by: Sharu Goel goelshar@amazon.com