Fix for cache reference leak warning (#3161)
Problem: Cache Reference Leaks
During the Alter function/procedure we are doing a sys-cache search without releasing it.
During the routines-definition, in one case we were not releasing the syscache search.

Fix:
a. To do sys-cache search via tuple and releasing it after the sys-cache search.
b. Also we are generating the logfile but the path that we were giving it github action workflow was not correct, updated that too.
c. Added a step in jdbc-test workflow which will compare the warnings in logfile with expected warnings

Task: BABEL-5551

Signed-off-by: Pranav Jain <pranavjc@amazon.com>