Add support for collation in partition functions and partition keys (#3556)
This commit adds the following changes to support collation in partition functions and partition keys:

1. Updated the CREATE PARTITION FUNCTION syntax to allow specifying the collation for the partition function
parameter. This allows users to create partition functions with a specific collation, instead of always using the
database default collation.

2. Modified the sys.babelfish_partition_function catalog table to include a new column 'input_parameter_collation' to
store the collation of the partition function parameter.

3. Updated the code that handles partition functions and partition keys to enforce that the collation of the partitioning
column must match the collation specified in the corresponding partition function. If they do not match, an error is
raised.

4. Added tests to verify the new partition function collation support and the validation of partition key collation against
the partition function collation.

Task: BABEL-1373
Signed-off-by: Sumit Jaiswal <sumiji@amazon.com>