Support syntax option INLINE = {ON | OFF} in CREATE FUNCTION (#3555) (#3606)
Currently, Babelfish throws syntax error when INLINE option is used in CREATE FUNCTION statements. The issue arises
from ANTLR's inability to recognise this token due to missing grammar in CREATE FUNCTION. To resolve this, we added
the INLINE = (ON | OFF) grammar to ANTLR and implemented the logic to ignore it during parsing time.

Task: BABEL-3719

Signed-off-by: Roshan Kanwar <rskanwar@amazon.com>