Remove redundant convert functions from string datatype to datetime. (#3619)
We have added helper functions to convert from string datatypes to datetime (in a previous [PR](https://github.com/babelfish-for-postgresql/babelfish_extensions/pull/3598)). We have now created a helper function with an 'anyelement' argument, keeping the rest of the code the same. This allows us to create a helper for convert functions from varbinary, while removing the helper functions for converting from string datatypes to datetime.

Task: BABEL-3070
Signed-off-by: yashneet vinayak <yashneet@amazon.com>

---------

Signed-off-by: yashneet vinayak <yashneet@amazon.com>
Co-authored-by: yashneet vinayak <yashneet@amazon.com>