Support column_list in CONTAINS clause (#3587)
Previously in Babelfish, the feature of fulltext search for simple terms in CONTAINS clause was not supported for multiple columns. The TSQL syntax of Fulltext search for simple terms in CONTAINS clause over multiple columns was unidentifiable over TDS endpoint of Babelfish and would throw syntax error. To fix this, we have created a new grammar to support the TSQL multiple column Fulltext search syntax for simple terms in CONTAINS clause over TDS endpoint of Babelfish. The support for multiple column fulltext search over CONTAINS clause is extended with validation check over columns if they are a part of existing fulltext index. The fulltext index can be created over 32 columns for a single table.

TASK: BABEL-4239

Signed-off-by: Jaspal Singh ijaspals@amazon.com