Support prefix term grammar for CONTAINS clause (#3588)
* Support prefix term grammar for CONTAINS clause

Previously in Babelfish, the feature of fulltext search for prefix term
as search string in CONTAINS clause was not supported. Prefix term over TDS endpoint
would throw unsupported error. To fix this, helper functions are created to
support the grammar of prefix term to provide the intended result.

TASK: BABEL-4380
Signed-off-by: Jaspal Singh ijaspals@amazon.com