Fix loading of data for TEXT column (NULL allowed) if primary key is defined using non-default collation from PG endpoint for logical database (#3824)
If the babelfish table has a column whose type is anything other than text/ntext (eg: varchar, nvarchar etc) and the collation of the column is NOT same as babelfish database/server level collation and if the query is executed from PSQL endpoint, it will throw error saying it can not decide which collation to use.

If the same query is executed from TDS endpoint, it runs successfully.

This is because when the query is executed from PSQL endpoint -- the collation of varchar is server level collation. During resolution of collation at the parent node, the collation of left and right operand does NOT match and none of their collation matches with CLUSTER_COLLATION_OID (this returns DEFAULT_COLLATION_OID as the dialect is PG)

The query is working from TDS endpoint -- the collation of varchar is database level collation. During resolution of collation at the parent node, the collation of left and right operand does NOT match and collation of right operand matches with CLUSTER_COLLATION_OID (this returns database level collation as the dialect is TSQL)

This commit handles the case when the connection is from PG endpoint and psql_logical_babelfish_db_name GUC is set -- the case for logical/TSQL database. If both the aforementioned conditions are met, we will return babelfish server level collation irrespective of dialect (initially it was only for TDS connection).

Task: BABEL-5077

Signed-off-by: Shameem Ahmed shmeeh@amazon.com