Rectified sys.database_principals view which was missing SID (#3730)
* Rectified sys.database_principals view which was missing SID

sys.database_principals.sid acts as a foreign key and links to sys.server_principals.sid
This is how SQL users are linked to their respective logins. Certain database_principals like db roles will not have a corresponding mapping in server_principals, hence they have a SID which is generated by default by sql server at creation.
In this view, we want to map the users (the defaults like dbo as well) to the server_principals. The owner of the database is the login which is linked to 'dbo'. Hence, we will be mapping the owner of the current db to 'dbo'. Other user-defined logins are already part of the view currently.

Task: BABEL-5789

Signed-off-by: Ayush Shah <ayushdsh@amazon.com>

---------

Signed-off-by: Ayush Shah <ayushdsh@amazon.com>