Adding operator class between numeric and fixeddecimal to support index scan clause (#3970)
Earlier in Babelfish, index scan would not be chosen by the optimizer when filtering between numeric and money/smallmoney/fixeddecimal types if the index was created on a numeric column.

This commit fixes this issue by adding a new operator class as part of the numeric operator family which allows the optimizer to utilize indexes on numeric columns when filtering with money/smallmoney/fixeddecimal values, improving query performance in these scenarios.

Engine PR : babelfish-for-postgresql/postgresql_modified_for_babelfish#605

Issues Resolved
BABEL-5959

Signed-off-by: Tanya Gupta tanyagp@amazon.com